+++ prisma/migrations/20260911073425_add_claimed_to_ledger/migration.sql
@@ -1,0 +1,2 @@
+-- AlterTable
+ALTER TABLE "ledgers" ADD COLUMN     "claimed" BOOLEAN NOT NULL DEFAULT false;